+++ assets/js/index.js
@@ -376,6 +376,50 @@ mheadleythemeapp.primaryMenu = {
 	}
 }; // mheadleythemeapp.primaryMenu
 
+
+
+
+/*	-----------------------------------------------------------------------------------------------
+	Comments and Ratings
+--------------------------------------------------------------------------------------------------- */
+
+mheadleythemeapp.commentRating = {
+
+	init: function() {
+		this.starClicks();
+	},
+
+	// The focusMenuWithChildren() function implements Keyboard Navigation in the Primary Menu
+	// by adding the '.focus' class to all 'li.menu-item-has-children' when the focus is on the 'a' element.
+	starClicks: function() {
+		// Get all the link elements within the primary menu.
+		var star, i, len,
+			ratingList = document.querySelector( '.ratings-container.interactive-rating .ratings-list' );
+
+		if ( ! ratingList ) {
+			return false;
+		}
+
+		star = ratingList.getElementsByTagName( 'input' );
+
+		// Each time a menu link is focused or blurred, toggle focus.
+		for ( i = 0, len = star.length; i < len; i++ ) {
+			star[i].addEventListener( 'change', setStarClass, true );
+		}
+
+		//Sets or removes the .focus class on an element.
+		function setStarClass(e) {
+			var self = this;
+      console.log(e);
+      if(!e.target){
+        return false;
+      }
+      ratingList.setAttribute("class", "ratings-list selected-rating-"+e.target.value);
+		}
+	}
+}; // mheadleythemeapp.comments
+
+
 /*	-----------------------------------------------------------------------------------------------
 	Toggles
 --------------------------------------------------------------------------------------------------- */
@@ -1529,6 +1573,7 @@ mheadleythemeappDomReady( function() {
 	mheadleythemeapp.coverModals.init();	// Handle cover modals
 	mheadleythemeapp.intrinsicRatioVideos.init();	// Retain aspect ratio of videos on window resize
 	mheadleythemeapp.primaryMenu.init();	// Primary Menu
+	mheadleythemeapp.commentRating.init();	// Comments
   mheadleythemeapp.touchEnabled.init();	// Add class to body if device is touch-enabled
   mheadleythemeapp.scrollReactor.init(); // adds scroll reactions and parallax effects
   mheadleythemeapp.scrollSegment.init(); //segment declarations and adding funtionality